# improve pitched legibility of labels that follow lines