… src/data/buffer_group.js
… src/render/draw_symbol.js
… src/render/painter.js
… src/render/vertex_array_object.js
… src/shaders/symbol_icon.vertex.glsl
… src/shaders/symbol_sdf.vertex.glsl
… src/symbol/collision_box.js
… src/symbol/collision_feature.js
… src/symbol/collision_tile.js
… src/symbol/projection.js
… src/symbol/quads.js
… src/symbol/symbol_size.js
… src/util/struct_array.js
… test/integration/query-tests/symbol-features-in/tilted-outside/style.json
… test/integration/render-tests/debug/collision-pitched-wrapped/expected.png
… test/integration/render-tests/debug/collision-pitched/expected.png
… test/integration/render-tests/icon-text-fit/placement-line/expected.png
… test/integration/render-tests/symbol-placement/point-polygon/expected.png
… test/integration/render-tests/symbol-spacing/line-close/expected.png
… test/integration/render-tests/symbol-spacing/line-overscaled/expected.png
… test/integration/render-tests/text-keep-upright/line-placement-false/expected.png
… test/integration/render-tests/text-keep-upright/line-placement-true-offset/expected.png
… test/integration/render-tests/text-keep-upright/line-placement-true-offset/style.json
… test/integration/render-tests/text-keep-upright/line-placement-true-rotated/expected.png
… test/integration/render-tests/text-keep-upright/line-placement-true/expected.png
… test/integration/render-tests/text-max-angle/literal/expected.png
… test/integration/render-tests/text-pitch-alignment/auto-text-rotation-alignment-map/expected.png
… test/integration/render-tests/text-pitch-alignment/map-text-depthtest/expected.png
… test/integration/render-tests/text-pitch-alignment/map-text-rotation-alignment-map/expected.png
… test/integration/render-tests/text-pitch-alignment/viewport-text-depthtest/expected.png