fixes #4718

This improves legibility of labels that follow lines in pitched views. The previous approach used the limited information in the shader to calculate put the glyph in approximatelyright place. The new approach does this more accurately by doing it on the cpu where we have access to the entire line geometry.

The core part of the pr is `src/symbol/projection.js` which does two things:
- calculates the matrices used for projecting symbols
- the find-the-correct-point-along-the-line-for-this-glyph calculations

See the comment in `src/symbol/projection.js` for an overview of the new approach.

This pr also gets rid of the the [sliding-glyphs-for-each-segment-that-get-toggled-on-and-off](https://camo.githubusercontent.com/5e1f9d174172207ae5a840dfae2ffec8b420661a/68747470733a2f2f662e636c6f75642e6769746875622e636f6d2f6173736574732f313432313635322f313535343730382f62373964623831342d346534622d313165332d393864612d3438646233336438353463652e676966) approach, which simplifies things a lot.

This fixes:
- garbled text near sharp corners
- upside down labels caused by perspective sloping
- glyph stretching and squishing
- letter spacing with pitch-scaling

### before / after

<img width="661" alt="screen shot 2017-06-02 at 6 25 10 pm" src="https://cloud.githubusercontent.com/assets/1421652/26747022/267acdea-47c1-11e7-84a8-f346314d9d65.png">
<img width="671" alt="screen shot 2017-06-02 at 6 23 44 pm" src="https://cloud.githubusercontent.com/assets/1421652/26747023/26e5e6e8-47c1-11e7-94df-5051e665767b.png">
<img width="447" alt="screen shot 2017-06-02 at 6 36 43 pm" src="https://cloud.githubusercontent.com/assets/1421652/26747260/9b85c51c-47c2-11e7-87aa-71f0258f9cd1.png">
<img width="446" alt="screen shot 2017-06-02 at 6 37 00 pm" src="https://cloud.githubusercontent.com/assets/1421652/26747261/9b8b73c2-47c2-11e7-8807-607cbe0d524f.png">
<img width="434" alt="screen shot 2017-06-02 at 6 44 23 pm" src="https://cloud.githubusercontent.com/assets/1421652/26747419/971addae-47c3-11e7-9e64-28e4c421a2ca.png">
<img width="435" alt="screen shot 2017-06-02 at 6 43 25 pm" src="https://cloud.githubusercontent.com/assets/1421652/26747418/9714ff38-47c3-11e7-896c-da9688377e2e.png">






## Benchmarks
Frame duration is about 1ms slower in benchmarks. In full page maps it's about 1-2 ms slower, and a bit more in pitched full page maps.

benchmark | master 943dc39 | along-line-label-legibility b5b6f9a
--- | --- | ---
**map-load** | 243 ms  | 138 ms 
**style-load** | 122 ms  | 119 ms 
**buffer** | 1,075 ms  | 1,069 ms 
**fps** | 60 fps  | 60 fps 
**frame-duration** | **4.2 ms**, 0% > 16ms  | **5.2 ms, 0% > 16ms**
**query-point** | 1.55 ms  | 1.41 ms 
**query-box** | 73.66 ms  | 73.58 ms 
**geojson-setdata-small** | 12 ms  | 5 ms 
**geojson-setdata-large** | 137 ms  | 130 ms 



## Launch Checklist

<!-- Thanks for the PR! Feel free to add or remove items from the checklist. -->

 - [x] briefly describe the changes in this PR
 - [ ] write tests for all new functionality
 - [x] post benchmark scores
 - [x] manually test the debug page

@ChrisLoer @kkaefer @anandthakker