@ansis 2017-06-02
and a couple wip shots:
<img width="773" alt="screen shot 2017-05-12 at 11 17 33 am" src="https://cloud.githubusercontent.com/assets/1421652/26747811/706443aa-47c6-11e7-82b0-2dea446ba431.png">
<img width="776" alt="screen shot 2017-05-12 at 11 45 20 am" src="https://cloud.githubusercontent.com/assets/1421652/26747821/7eae24f8-47c6-11e7-9416-baf83cacfbff.png">
<img width="777" alt="screen shot 2017-05-12 at 12 50 14 pm" src="https://cloud.githubusercontent.com/assets/1421652/26747835/90b6fb52-47c6-11e7-95b2-596ece3ef9bd.png">