@ChrisLoer 2017-06-02
This looks SOO good!

I won't have a chance to get into too much review until Monday, but to give you an idea for how the old `max-camera-distance` filter might still be relevant, look at:

`14.27/34.004/-118.4845/-46.6/60` with viewport aligned road labels:

![screenshot 2017-06-02 16 47 34](https://cloud.githubusercontent.com/assets/375121/26748515/3ecb367c-47b3-11e7-8309-3bc9159a1486.png)