@ChrisLoer 2017-06-05
Example of broken collision detection for vertically-oriented viewport-pitch-aligned road labels:

![image](https://cloud.githubusercontent.com/assets/375121/26798423/3c05df44-49e5-11e7-8248-f00fedeeae14.png)