@ansis 2017-06-06
@ChrisLoer yeah, the collision boxes are a problem. I just tried adding an adjustment to `maxScale` to account for this and it seems like it mostly works: https://github.com/mapbox/mapbox-gl-js/pull/4781/commits/6a789ecff129cae92d107cbe85d14a537340d206. Do you think this is good enough to last until we rewrite this?