@ChrisLoer 2017-06-06
This is a nit, and it could go in another PR, but maybe it would be better if we detected the orientation of a line label based on the beginning and the end of the label instead of the orientation of the middle line segment? The current approach can lead to labels that look "upside down" even though technically the segment their anchor is on is "right side up":

![image](https://user-images.githubusercontent.com/375121/26845174-8130bbdc-4aaa-11e7-886d-e8c8933fc766.png)