@ChrisLoer 2017-06-07
With an eye on https://github.com/mapbox/mapbox-gl-native/pull/9009#issuecomment-306276107, we should try to see if we can stay under the 8 vertex attribute limit when we make these changes on the JS side.

> since we should move towards treating `a_size` as a regular data-driven property we might not want to pack into it.

Do we have a choice? On this PR `a_data` is using 40 bits and `a_size` is using 32 bits, and I believe the only extra space is 16 bits in `a_projected_pos`. If we keep `a_size` and try to get rid of `a_data`, we'd have 40 - 16 -> 24 bits to pack in somewhere else. We can drop a few bits of precision from the offset in `a_pos_offset` (dropping from 1/64 precision to 1/8 saves 6 bits and doesn't look too noticeably different although the render tests do change), and a few bits from the anchor position. You could also go back to dropping 8 bits of precision from `segment_angle` (but that'd be a bummer, because the low precision angles meant that labels sometimes didn't quite line up with their lines).