@ansis 2017-06-12
I just rebased on master which includes the recent `*-offset` fix (https://github.com/mapbox/mapbox-gl-js/pull/4779). The new implementation is [here](https://github.com/mapbox/mapbox-gl-js/blob/75f0765dfded4afbb893952449fc9d76923d2f35/src/symbol/projection.js#L251) for alongLine labels and [here](https://github.com/mapbox/mapbox-gl-js/blob/75f0765dfded4afbb893952449fc9d76923d2f35/src/symbol/quads.js#L157) for horizontal labels.

> Out of curiosity, why did you pack the label angle into the third part of a_projected_pos instead of just adding a fourth item to the vector? Do we think there's a performance benefit? I was a little sad to see the increased label angle precision go away (and a little surprised that text-writing-mode/chinese didn't have to be reverted after the angle change)

I'm not sure, I guess the impulse was to just pack everything as tightly as possible instead of using an extra 4 bytes. Should I switch to using a fourth item?

my remaining items:
- fix the `x` component of `text-offset` for labels that follow lines
- fix failing query test