@ansis 2017-06-12
@ChrisLoer I also made a query test less sensitive to avoid a failure. The test *shouldn't* be less sensitive but I don't think we can fix this properly before working on all the collision stuff.